Extract forwarder package
