--- core/src/main/java/com/novoda/noplayer/exoplayer/ExoPlayerBinder.java
@@ -1,15 +1,0 @@
-package com.novoda.noplayer.exoplayer;
-
-public class ExoPlayerBinder {
-
-    private final ExoPlayerTwoFacade facade;
-
-    ExoPlayerBinder(ExoPlayerTwoFacade facade) {
-        this.facade = facade;
-    }
-
-    public void bind(ExoPlayerForwarder exoPlayerForwarder) {
-        exoPlayerForwarder.bind(facade);
-    }
-
-}
