+++ core/src/main/java/com/novoda/noplayer/exoplayer/forwarder/Bitrate.java
@@ -1,4 +1,4 @@
-package com.novoda.noplayer.exoplayer;
+package com.novoda.noplayer.exoplayer.forwarder;
 
 public final class Bitrate {
