+++ core/src/main/java/com/novoda/noplayer/exoplayer/forwarder/BitrateForwarder.java
@@ -1,4 +1,4 @@
-package com.novoda.noplayer.exoplayer;
+package com.novoda.noplayer.exoplayer.forwarder;
 
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.Format;
