+++ core/src/main/java/com/novoda/noplayer/exoplayer/forwarder/BufferStateForwarder.java
@@ -1,4 +1,4 @@
-package com.novoda.noplayer.exoplayer;
+package com.novoda.noplayer.exoplayer.forwarder;
 
 import com.google.android.exoplayer2.ExoPlaybackException;
 import com.google.android.exoplayer2.ExoPlayer;
@@ -6,28 +6,23 @@ import com.google.android.exoplayer2.PlaybackParameters;
 import com.google.android.exoplayer2.Timeline;
 import com.google.android.exoplayer2.source.TrackGroupArray;
 import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
-import com.novoda.noplayer.PlayerState;
-import com.novoda.noplayer.listeners.PreparedListeners;
+import com.novoda.noplayer.listeners.BufferStateListeners;
 
-class OnPrepareForwarder implements ExoPlayer.EventListener {
+class BufferStateForwarder implements ExoPlayer.EventListener {
 
-    private final PreparedListeners preparedListeners;
-    private final PlayerState playerState;
+    private final BufferStateListeners bufferStateListeners;
 
-    OnPrepareForwarder(PreparedListeners preparedListeners, PlayerState playerState) {
-        this.preparedListeners = preparedListeners;
-        this.playerState = playerState;
+    BufferStateForwarder(BufferStateListeners bufferStateListeners) {
+        this.bufferStateListeners = bufferStateListeners;
     }
 
     @Override
     public void onPlayerStateChanged(boolean playWhenReady, int playbackState) {
-        if (isReady(playbackState)) {
-            preparedListeners.onPrepared(playerState);
+        if (playbackState == ExoPlayer.STATE_BUFFERING) {
+            bufferStateListeners.onBufferStarted();
+        } else if (playbackState == ExoPlayer.STATE_READY) {
+            bufferStateListeners.onBufferCompleted();
         }
-    }
-
-    private boolean isReady(int playbackState) {
-        return playbackState == ExoPlayer.STATE_READY;
     }
 
     @Override
