+++ core/src/main/java/com/novoda/noplayer/exoplayer/forwarder/ExoPlayerErrorFactory.java
@@ -1,4 +1,4 @@
-package com.novoda.noplayer.exoplayer;
+package com.novoda.noplayer.exoplayer.forwarder;
 
 import android.media.MediaCodec;
