+++ core/src/main/java/com/novoda/noplayer/exoplayer/forwarder/InternalErrorListener.java
@@ -1,6 +1,7 @@
-package com.novoda.noplayer.exoplayer;
+package com.novoda.noplayer.exoplayer.forwarder;
 
 import com.google.android.exoplayer2.ExoPlaybackException;
+import com.novoda.noplayer.exoplayer.forwarder.ExoPlayerForwarder;
 
 import java.io.IOException;
