+++ core/src/main/java/com/novoda/noplayer/exoplayer/forwarder/VideoSizeChangedForwarder.java
@@ -1,4 +1,4 @@
-package com.novoda.noplayer.exoplayer;
+package com.novoda.noplayer.exoplayer.forwarder;
 
 import android.view.Surface;
