+++ core/src/main/java/com/novoda/noplayer/listeners/BitrateChangedListeners.java
@@ -1,7 +1,7 @@
 package com.novoda.noplayer.listeners;
 
 import com.novoda.noplayer.Player;
-import com.novoda.noplayer.exoplayer.Bitrate;
+import com.novoda.noplayer.exoplayer.forwarder.Bitrate;
 
 import java.util.Set;
 import java.util.concurrent.CopyOnWriteArraySet;
